refactored application.properties added classes and mathods to recieve and save 'photo' in DB

+++ src/main/java/com/staj/staj/common_jpa/entity/AppPhoto.java
@@ -16,5 +16,5 @@ public class AppPhoto {
     private String telegramFieldId;
     @OneToOne
     private BinaryContent binaryContent;//ссылка на объект BinaryContent
-    private Long fileSize;
+    private  Integer fileSize;
 }
